# docs/swagger.yaml
definitions:
...
  httputils.ErrorResponse:
    properties:
      error:
        example: failed to parse query value 'stars-from'
        type: string
      message:
        example: failed to find hotels
        type: string
      success:
        example: false
        type: boolean
    type: object
  httputils.SuccessResponse:
    properties:
      data: {}
      message:
        example: successfully found hotels
        type: string
      success:
        example: true
        type: boolean
    type: object
  rent.Rent:
    properties:
      date_from:
        type: integer
      date_to:
        type: integer
      hotel_room_id:
        type: string
      id:
        type: string
      renter_id:
        type: string
    type: object
  user.User:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/user.UserRole'
    type: object
  user.UserRole:
    enum:
    - admin
    - user
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleUser
info:
  contact: {}
paths:
...
  /api/hotel-room/rent:
    post:
      consumes:
      - application/json
      description: Бронирует номер отеля на заданные даты. Дата выезда считается свободной
        для другого бронирования. 'days' нельзя использовать вместе с 'date-to'. Требуется
        авторизация
      parameters:
      - description: ID номера отеля
        in: query
        name: id
        required: true
        type: string
      - description: Дата заселения (прим. 2016-10-06, год-месяц-день)
        in: query
        name: date-from
        required: true
        type: string
      - description: Дата выезда (прим. 2016-10-06, год-месяц-день) (по выбору)
        in: query
        name: date-to
        type: string
      - description: Количество дней (по выбору)
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/rent.Rent'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Забронировать номер отеля
      tags:
      - Номер отеля
  /api/hotel-room/unrent:
    post:
      consumes:
      - application/json
      description: Отменяет бронь номер отеля по ID брони. Админ может удалить бронь
        у любого пользователя. Требуется авторизация
      parameters:
      - description: ID брони
        in: query
        name: rent-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Отменить бронь номера отеля
      tags:
      - Номер отеля
...
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Вход в аккаунт по email и паролю
      parameters:
      - description: Email
        in: query
        name: email
        required: true
        type: string
      - description: Пароль
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Войти в аккаунт
      tags:
      - Пользователь
  /api/user/register:
    post:
      consumes:
      - application/json
      description: Регистрирует пользователя по email, имени и паролю
      parameters:
      - description: Имя пользователя
        in: query
        name: name
        required: true
        type: string
      - description: Email пользователя
        in: query
        name: email
        required: true
        type: string
      - description: Пароль пользователя
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Зарегистрироваться
      tags:
      - Пользователь
  /api/user/rename:
    put:
      consumes:
      - application/json
      description: Изменяет имя пользователя. Требуется авторизация
      parameters:
      - description: Новое имя пользователя
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Изменить имя
      tags:
      - Пользователь
swagger: "2.0"